Explicity set black on paths
Currently paths which are black do not have a color set. This defaults to black when viewing the SVG in isolation.
However, when embedded in a website that sets a different default fill colour for SVGs it makes the black paths show with the wrong colour.

As those country flag colours should be explicitly black I've set the fill color on all paths that didn't have it so they are unaffected by the webpage default styles.